migrateファイル編集2

+++ db/migrate/20220703052128_create_posts.rb
@@ -1,7 +1,7 @@
 class CreatePosts < ActiveRecord::Migration[6.1]
   def change
     create_table :posts do |t|
-      t.references :user, type: :integer, null: false, foreign_key: true
+      t.references :user, type: :bigint, null: false, foreign_key: true
       t.string :body, null: false
       t.decimal :lat
       t.decimal :long
