+++ db/migrate/20220703052540_create_ranks.rb
@@ -1,8 +1,8 @@
 class CreateRanks < ActiveRecord::Migration[6.1]
   def change
     create_table :ranks do |t|
-      t.references :user, type: :integer, null: false, foreign_key: true
-      t.references :post, type: :integer, null: false, foreign_key: true
+      t.references :user, type: :bigint, null: false, foreign_key: true
+      t.references :post, type: :bigint, null: false, foreign_key: true
       t.integer :rank
 
       t.timestamps
