+++ db/migrate/20220703054235_create_post_tags.rb
@@ -1,8 +1,8 @@
 class CreatePostTags < ActiveRecord::Migration[6.1]
   def change
     create_table :post_tags do |t|
-      t.references :post, type: :integer, null: false, foreign_key: true
-      t.references :tag, type: :integer, null: false, foreign_key: true
+      t.references :post, type: :bigint, null: false, foreign_key: true
+      t.references :tag, type: :bigint, null: false, foreign_key: true
 
       t.timestamps
     end
