+++ db/migrate/20220703054728_create_favorites.rb
@@ -1,8 +1,8 @@
 class CreateFavorites < ActiveRecord::Migration[6.1]
   def change
     create_table :favorites do |t|
-      t.references :user, type: :integer, null: false, foreign_key: true
-      t.references :post, type: :integer, null: false, foreign_key: true
+      t.references :user, type: :bigint, null: false, foreign_key: true
+      t.references :post, type: :bigint, null: false, foreign_key: true
 
       t.timestamps
     end
